+++ src/types/fragments.ts
@@ -1,7 +1,11 @@
-import type { OrderingOptions } from "./state";
+import type { OrderingOptions } from './state';
 
 export interface FragmentsOptions extends OrderingOptions {}
 
+export interface FragmentLoader {
+  (name: string, parameters: any, options?: FragmentsOptions): Promise<string>;
+}
+
 export interface FragmentsService {
-  load(name: string, parameters: any, options?: FragmentsOptions): Promise<string>;
+  load: FragmentLoader;
 }