+++ src/types/part.ts
@@ -1,0 +1,5 @@
+import type { Document } from 'domhandler';
+
+export interface PartService {
+  getReplacement(document: Document, attribs: Record<string, string>): Promise<string>;
+}